创建项目

+++ jb4dc-base-root/jb4dc-base-service/src/main/java/com/jb4dc/base/service/aspect/ClientCallRemoteCache.java
@@ -10,5 +10,5 @@ import java.lang.annotation.Target;
 @Target(ElementType.METHOD)
 @Retention(RetentionPolicy.RUNTIME)
 public @interface ClientCallRemoteCache {
-    int expirationTimeSeconds() default 300;
+    long expirationTimeSeconds() default 300;
 }
